refactor: move postPerformance to project.controller.js
